Moved STD-like functions to Array.h
